refactor: chat 패키지 코드 품질 개선 - suspend 전환, validation, DTO 분리
- runBlocking 제거 → Controller/Service를 suspend로 전환하여 스레드 블로킹 방지
- ChatRequest.message에 @field:NotBlank, @field:Schema 추가 및 @Valid 적용
- IllegalArgumentException → CustomException(SafersErrorCode.NOT_FOUND_SITE) 컨벤션 통일
- ScreenMeta, ScreenCache를 chat/dto/ScreenDto.kt로 분리
- CORS 허용 origin 추가
