+++ apps/safers/src/main/kotlin/com/pluxity/safers/chat/service/ChatHistoryStore.kt
@@ -2,6 +2,8 @@ package com.pluxity.safers.chat.service
 
 import com.pluxity.safers.chat.dto.ChatResponse
 import com.pluxity.safers.chat.dto.QueryAction
+import com.pluxity.safers.chat.dto.ScreenCache
+import com.pluxity.safers.chat.dto.ScreenMeta
 import com.pluxity.safers.llm.LlmClient
 import com.pluxity.safers.llm.dto.Message
 import io.github.oshai.kotlinlogging.KotlinLogging
@@ -10,26 +12,6 @@ import org.springframework.stereotype.Component
 import java.time.Duration
 
 private val log = KotlinLogging.logger {}
-
-/**
- * 캐시된 화면 메타데이터 (프롬프트에 포함용)
- */
-data class ScreenMeta(
-    val ref: String,
-    val summary: String,
-    val siteIds: List<Long>,
-    val targets: List<String>,
-    val actionIds: List<String>,
-)
-
-/**
- * 캐시된 화면 전체 (복원용)
- */
-data class ScreenCache(
-    val meta: ScreenMeta,
-    val actions: List<QueryAction>,
-    val response: ChatResponse,
-)
 
 @Component
 class ChatHistoryStore(
